# feat: improve azure latency by reusing http clients