--- a/Cargo.toml
+++ b/Cargo.toml
@@ -47,7 +47,7 @@ reqwest = { version = "^0.12", default-features = false, features = [
     "json",
     "rustls-tls",
 ] }
-iceberg = { git = "https://github.com/hansetag/iceberg-rust.git", rev = "2b6263a6c998df75444f7feb54bd37144354a071", features = [
+iceberg = { git = "https://github.com/hansetag/iceberg-rust.git", rev = "acdfca1dd68e2f7bb38afd64f77648102d135d4f", features = [
     "storage-all",
 ] }
 typed-builder = "^0.20.0"