reuses http clients against azure and reduces fetching sas from 800ms to 200ms for subsequent calls, also makes creation of table faster, down from 1700ms to 1200ms

#359